S08P22C205-378 set cookie -ing

+++ jobtalkbackend/src/main/java/com/ssafy/jobtalkbackend/jwt/JwtTokenProvider.java
@@ -32,10 +32,10 @@ public class JwtTokenProvider {
     private String secret_key;
 
     @Value("${jwt.token-validity-in-milliseconds}")
-    private long accessTokenValidityTime;
+    private int accessTokenValidityTime;
 
     @Value("${jwt.refresh-token-validity-in-milliseconds}")
-    private long refreshTokenValidityTime;
+    private int refreshTokenValidityTime;
 
     @Autowired
     private UserDetailsService userDetailsService;
